requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
aiohttp==3.9.3
requests==2.31.0

# Slack
slack-bolt==1.18.0
slack-sdk==3.26.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Celery
celery==5.3.6
redis==5.0.1

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
prometheus-client==0.19.0 